+++ internal/features/advert/repo/postgres/get_by_id.go
@@ -10,7 +10,7 @@ import (
 
 func (r *Repo) GetByID(
 	ctx context.Context,
-	id int,
+	id int64,
 ) (advertentity.Advert, error) {
 	ctx, cancel := context.WithTimeout(ctx, r.pool.OpTimeout())
 	defer cancel()
