+++ internal/features/category/controller/response.go
@@ -5,8 +5,8 @@ import (
 )
 
 type categoryResponse struct {
-	ID       int    `json:"id"`
-	ParentID *int   `json:"parent_id"`
+	ID       int64  `json:"id"`
+	ParentID *int64 `json:"parent_id"`
 	Name     string `json:"name"`
 }
